IOS-7684: provide imprecise and instantaneous node count

+++ Modules/Domain/MEGADomain/Sources/MEGADomain/RepositoryProtocol/Account/AccountRepositoryProtocol.swift
@@ -8,7 +8,7 @@ public protocol AccountRepositoryProtocol: RepositoryProtocol {
     var bandwidthOverquotaDelay: Int64 { get }
     func isLoggedIn() -> Bool
     func contacts() -> [UserEntity]
-    func totalNodesCount() -> UInt
+    func totalNodesCount() -> UInt64
     func getMyChatFilesFolder(completion: @escaping (Result<NodeEntity, AccountErrorEntity>) -> Void)
     func upgradeSecurity() async throws -> Bool
     func incomingContactsRequestsCount() -> Int
